requirements.txt
iexfinance==0.5.0
matplotlib==3.3.4
pandas==1.1.5
opencv-python-headless<4.3
discord.py


